rm energyDensity.c
