+++ src/makefile
@@ -82,7 +82,7 @@ OBJSC = main.o initialization.o readfiles.o atomdata.o parallelization.o relax.o
         occupation.o gradVecRoutines.o gradVecRoutinesKpt.o nlocVecRoutines.o     \
         hamiltonianVecRoutines.o lapVecRoutines.o lapVecRoutinesKpt.o \
         linearSolver.o mixing.o exchangeCorrelation.o eigenSolver.o eigenSolverKpt.o energy.o      \
-        forces.o stress.o pressure.o finalization.o spinOrbitCoupling.o printing.o energyDensity.o \
+        forces.o stress.o pressure.o finalization.o spinOrbitCoupling.o printing.o \
         linearAlgebra.o \
         xc/vdW/d3/d3correction.o xc/vdW/d3/d3findR0ab.o xc/vdW/d3/d3copyC6.o                       \
         xc/vdW/d3/d3initialization.o xc/vdW/d3/d3finalization.o xc/vdW/d3/d3forceStress.o          \
